bgpd: fix SRv6 L3VPN route leak to VPNv4 and VPNv6 SID nexthop validity
When a IPv4 or IPv6 route belonging to a VRF is leaked to VPNv4 or VPNv6 using
SRv6, the locally assigned SID never contains nexthops (bnc->nexthop_num > 0)
and never passes bgp_isvalid_nexthop().

This change updates leak_update_nexthop_valid() to assume the nexthop is valid
if SRv6 is enabled and a SID is allocated. It will still fail if SRv6 is enabled
and neither a srv6_l3vpn or a srv6_vpn SID is assigned.