lib: logging: fix CLI log facility configuration
When f90bd85 was introduced, it changed the CLI output
of log facility configuration from:

	log facility local7

to the full YANG identity ref:

	log facility ietf-syslog-types:local7

However, this is invalid syntax for the CLI and causes tooling
like frr-reload.py to break.

This commit corrects output back to the original format while
not modifying any internal references.

Signed-off-by: Jonathan Voss <jvoss@onvox.net>